refactor(agents): split system.ts into per-purpose agent files
system.ts had grown to ~1270 lines (16 hand-written agents + two per-LV
generators + the registry) and become a constant merge-conflict hotspot —
concurrent agent work routinely collides in it. Split the definitions into
focused sibling files (coreAgents, oeffentlichkeitsarbeitAgents, personaAgents,
lvPrAgents, lvBuergerAgents) so unrelated agent changes touch different files,
leaving system.ts as a ~50-line aggregator that assembles SYSTEM_AGENTS and
resolves identifiers.

Behaviour-preserving: agent definitions are moved byte-for-byte and reassembled
in the same order, so SYSTEM_AGENTS is unchanged. The literal-typed SystemAgentId
is now derived from the three base arrays plus the spec arrays. Public exports
are untouched, so the internal consumers (slug/audience/mcpProjection/index) need
no changes.
